+++ src/dm.rs
@@ -18,18 +18,71 @@ use std::time::Duration;
 use nix::libc::ioctl as nix_ioctl;
 use nix::libc::c_ulong;
 
-
-use super::consts::{DM_NAME_LEN, DM_UUID_LEN, MIN_BUF_SIZE, DM_IOCTL, DmFlags, DM_CTL_PATH,
-                    DM_VERSION_MAJOR, DM_VERSION_MINOR, DM_VERSION_PATCHLEVEL, DM_READONLY,
-                    DM_SUSPEND, DM_PERSISTENT_DEV, DM_STATUS_TABLE, DM_BUFFER_FULL,
-                    DM_SKIP_LOCKFS, DM_NOFLUSH, DM_QUERY_INACTIVE_TABLE, DM_UUID, DM_DATA_OUT,
-                    DM_DEFERRED_REMOVE};
 use super::device::Device;
-use super::deviceinfo::DeviceInfo;
+use super::deviceinfo::{DM_NAME_LEN, DM_UUID_LEN, DeviceInfo};
 use super::dm_ioctl as dmi;
 use super::result::{DmError, DmResult, ErrorEnum};
 use super::types::{Sectors, TargetLine, TargetLineArg};
 use super::util::{align_to, slice_to_null};
+
+/// Indicator to send IOCTL to DM
+const DM_IOCTL: u8 = 0xfd;
+/// Control path for user space to pass IOCTL to kernel DM
+const DM_CTL_PATH: &'static str = "/dev/mapper/control";
+/// Major version
+const DM_VERSION_MAJOR: u32 = 4;
+/// Minor version
+const DM_VERSION_MINOR: u32 = 30;
+/// Patch level
+const DM_VERSION_PATCHLEVEL: u32 = 0;
+
+/// Start with a large buffer to make BUFFER_FULL rare. Libdm does this too.
+const MIN_BUF_SIZE: usize = 16 * 1024;
+
+bitflags! {
+    /// Flags used by devicemapper.
+    #[derive(Default)]
+    pub struct DmFlags: dmi::__u32 {
+        /// In: Device should be read-only.
+        /// Out: Device is read-only.
+        #[allow(identity_op)]
+        const DM_READONLY             = (1 << 0);
+        /// In: Device should be suspended.
+        /// Out: Device is suspended.
+        const DM_SUSPEND              = (1 << 1);
+        /// In: Use passed-in minor number.
+        const DM_PERSISTENT_DEV       = (1 << 3);
+        /// In: STATUS command returns table info instead of status.
+        const DM_STATUS_TABLE         = (1 << 4);
+        /// Out: Active table is present.
+        const DM_ACTIVE_PRESENT       = (1 << 5);
+        /// Out: Inactive table is present.
+        const DM_INACTIVE_PRESENT     = (1 << 6);
+        /// Out: Passed-in buffer was too small.
+        const DM_BUFFER_FULL          = (1 << 8);
+        /// Obsolete.
+        const DM_SKIP_BDGET           = (1 << 9);
+        /// In: Avoid freezing filesystem when suspending.
+        const DM_SKIP_LOCKFS          = (1 << 10);
+        /// In: Suspend without flushing queued I/Os.
+        const DM_NOFLUSH              = (1 << 11);
+        /// In: Query inactive table instead of active.
+        const DM_QUERY_INACTIVE_TABLE = (1 << 12);
+        /// Out: A uevent was generated, the caller may need to wait for it.
+        const DM_UEVENT_GENERATED     = (1 << 13);
+        /// In: Rename affects UUID field, not name field.
+        const DM_UUID                 = (1 << 14);
+        /// In: All buffers are wiped after use. Use when handling crypto keys.
+        const DM_SECURE_DATA          = (1 << 15);
+        /// Out: A message generated output data.
+        const DM_DATA_OUT             = (1 << 16);
+        /// In: Do not remove in-use devices.
+        /// Out: Device scheduled to be removed when closed.
+        const DM_DEFERRED_REMOVE      = (1 << 17);
+        /// Out: Device is suspended internally.
+        const DM_INTERNAL_SUSPEND     = (1 << 18);
+    }
+}
 
 /// Returns an error if value is unsuitable.
 fn dev_id_check(value: &str, max_allowed_chars: usize) -> DmResult<()> {
@@ -337,9 +390,7 @@ impl DM {
     /// # Example
     ///
     /// ```no_run
-    /// use devicemapper::{DM, DmName};
-    /// use devicemapper::consts::DmFlags;
-    ///
+    /// use devicemapper::{DM, DmFlags, DmName};
     ///
     /// let dm = DM::new().unwrap();
     ///
@@ -435,8 +486,7 @@ impl DM {
     /// # Example
     ///
     /// ```no_run
-    /// use devicemapper::{DM, DevId, DmName};
-    /// use devicemapper::consts::{DmFlags, DM_SUSPEND};
+    /// use devicemapper::{DM, DM_SUSPEND, DevId, DmFlags, DmName};
 
     /// let dm = DM::new().unwrap();
     ///
@@ -520,7 +570,6 @@ impl DM {
     ///
     /// ```no_run
     /// use devicemapper::{DM, DevId, DmName, Sectors};
-    /// use devicemapper::consts::DmFlags;
     /// let dm = DM::new().unwrap();
     ///
     /// // Create a 16MiB device (32768 512-byte sectors) that maps to /dev/sdb1
@@ -722,8 +771,7 @@ impl DM {
     /// # Example
     ///
     /// ```no_run
-    /// use devicemapper::{DM, DevId, DmName};
-    /// use devicemapper::consts::{DM_STATUS_TABLE, DmFlags};
+    /// use devicemapper::{DM, DM_STATUS_TABLE, DevId, DmFlags, DmName};
     /// let dm = DM::new().unwrap();
     ///
     /// let name = DmName::new("example-dev").expect("is valid DM name");
@@ -832,7 +880,6 @@ impl DM {
 mod tests {
 
     use {DevId, DM};
-    use consts::{DmFlags, DM_STATUS_TABLE};
 
     use super::*;
