+++ src/lib.rs
@@ -71,7 +71,7 @@ mod util;
 /// basic types (Bytes, Sectors, DataBlocks)
 mod types;
 /// shared constants
-pub mod consts;
+mod consts;
 /// Macros shared by device mapper devices.
 #[macro_use]
 mod shared_macros;
@@ -98,7 +98,17 @@ mod shared;
 mod loopbacked;
 
 
-pub use dm::{DM, DevId, DmName, DmNameBuf, DmUuid, DmUuidBuf};
+pub use consts::{IEC, SECTOR_SIZE};
+pub use dm::{DM, DevId, DmFlags, DmName, DmNameBuf, DmUuid, DmUuidBuf};
+
+// Export the DmFlags individually. In theory, they can be accessed by
+// DmFlags::DM_READONLY, but in practice this requires explicitly using
+// an obscure trait generated by the bitflags! macro.
+pub use dm::{DM_READONLY, DM_SUSPEND, DM_PERSISTENT_DEV, DM_STATUS_TABLE, DM_ACTIVE_PRESENT,
+             DM_INACTIVE_PRESENT, DM_BUFFER_FULL, DM_SKIP_BDGET, DM_SKIP_LOCKFS, DM_NOFLUSH,
+             DM_QUERY_INACTIVE_TABLE, DM_UEVENT_GENERATED, DM_UUID, DM_SECURE_DATA, DM_DATA_OUT,
+             DM_DEFERRED_REMOVE, DM_INTERNAL_SUSPEND};
+
 pub use device::Device;
 pub use lineardev::LinearDev;
 pub use result::{DmResult, DmError, ErrorEnum};
